Move clients building into own file
