+++ src/public/client.js
@@ -1,0 +1,30 @@
+import { ApolloClient } from 'apollo-client';
+import {
+  InMemoryCache,
+  IntrospectionFragmentMatcher,
+} from 'apollo-cache-inmemory';
+import { HttpLink } from 'apollo-link-http';
+import introspectionQueryResultData from './fragmentTypes.json';
+
+const client = new ApolloClient({
+  link: new HttpLink({
+    uri: 'https://platform-api.now.sh/graphql',
+  }),
+  cache: new InMemoryCache(),
+});
+
+const fragmentMatcher = new IntrospectionFragmentMatcher({
+  introspectionQueryResultData,
+});
+
+// TODO: This client uses the production GraphQL endpoint.
+// In the future, the whole app should use only one GraphQL endpoint.
+const client2 = new ApolloClient({
+  link: new HttpLink({
+    uri: 'https://api-beta-dot-open-targets-eu-dev.appspot.com/api/v4/graphql',
+  }),
+  cache: new InMemoryCache({ fragmentMatcher }),
+});
+
+export { client2 };
+export default client;
